test(ts): move ci bench util to shared package
