+++ crypto-ffi/bindings/js/packages/browser/benches/utils.ts
@@ -1,14 +1,12 @@
 import { browser } from "@wdio/globals";
 import type { Bench } from "tinybench";
 import { type CommitBundle } from "@wireapp/core-crypto/browser";
-import { isNumberObject } from "node:util/types";
-import { mkdir } from "node:fs/promises";
-import { writeFile } from "node:fs/promises";
 import {
     sharedSetup,
     type Helpers,
     type DeliveryService,
 } from "../shared/utils";
+import { logResults } from "../../shared/benches/utils";
 type ccModuleType = typeof import("@wireapp/core-crypto/browser");
 declare global {
     interface Window {
@@ -30,72 +28,6 @@ export async function setup() {
                 // @ts-expect-error TS2307: Cannot find module ./corecrypto.js or its corresponding type declarations.
                 await import("/node_modules/tinybench/dist/index.js");
         }
-    });
-}
-
-export type CustomBenchmarkEntry = {
-    name: string;
-    unit: string;
-    value: number;
-    range?: string;
-    extra?: string;
-};
-
-function parseMetric(metric?: string | number) {
-    if (!metric) return null;
-
-    if (isNumberObject(metric)) {
-        return {
-            metric,
-            range: undefined,
-        };
-    } else {
-        const [rawValue, rawRange] = metric
-            .split("±")
-            .map((part) => part.trim());
-
-        const value =
-            rawValue !== undefined ? Number.parseFloat(rawValue) : undefined;
-
-        if (Number.isNaN(value)) return null;
-
-        return {
-            value,
-            range: rawRange,
-        };
-    }
-}
-
-export function toCustomBenchmarkEntries(
-    benchmarkName: string | undefined,
-    rows: (Record<string, string | number | undefined> | null)[]
-): CustomBenchmarkEntry[] {
-    return rows.map((row) => {
-        const throughput = parseMetric(row?.["Throughput avg (ops/s)"]);
-        const extra = [
-            row?.["Latency avg (ns)"]
-                ? `Average Latency (ns): ${row["Latency avg (ns)"]}`
-                : null,
-            row?.["Latency med (ns)"]
-                ? `Median Latency (ns): ${row["Latency med (ns)"]}`
-                : null,
-            row?.["Throughput med (ops/s)"]
-                ? `Median Throughput (ops/s): ${row["Throughput med (ops/s)"]}`
-                : null,
-            row?.["Samples"] !== undefined
-                ? `Samples: ${row["Samples"]}`
-                : null,
-        ]
-            .filter(Boolean)
-            .join("\n");
-
-        return {
-            name: `${benchmarkName} - ${row?.["Task name"]}`,
-            unit: "ops/s",
-            value: throughput?.value ?? 0,
-            range: throughput?.range,
-            extra: extra || undefined,
-        };
     });
 }
 
@@ -122,16 +54,5 @@ export async function collectBenchmarkResults() {
         return { name: window.bench.name, table: window.bench.table() };
     });
 
-    console.log(results.name);
-    console.log(results.table);
-
-    if (!process.env["CI"]) return;
-
-    const customResults = toCustomBenchmarkEntries(results.name, results.table);
-
-    await mkdir("benches_result", { recursive: true });
-    await writeFile(
-        `benches_result/${results.name}.json`,
-        JSON.stringify(customResults, null, 2)
-    );
+    await logResults(results.name, results.table);
 }
